// react-module/src/openapi/models/atomic-product-definition-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductDefinitionBasicDto
 */
export interface AtomicProductDefinitionBasicDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductDefinitionBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductDefinitionBasicDto
     */
    title?: any;
    /**
     * 
     * @type {AtomicProductDefinitionBasicDto}
     * @memberof AtomicProductDefinitionBasicDto
     */
    parent?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductParameterBasicDto&gt;}
     * @memberof AtomicProductDefinitionBasicDto
     */
    parameters?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductBasicDto&gt;}
     * @memberof AtomicProductDefinitionBasicDto
     */
    atomicProducts?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductDefinitionBasicDto
     */
    category?: AtomicProductDefinitionBasicDtoCategoryEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicProductDefinitionBasicDtoCategoryEnum {
    BASE = 'BASE',
    INTERMEDIATE = 'INTERMEDIATE',
    FINAL = 'FINAL'
}