// react-module/src/openapi/models/storage-area-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface StorageAreaBasicDto
 */
export interface StorageAreaBasicDto {
    /**
     * 
     * @type {number}
     * @memberof StorageAreaBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof StorageAreaBasicDto
     */
    code?: any;
    /**
     * 
     * @type {WarehouseBasicDto}
     * @memberof StorageAreaBasicDto
     */
    warehouse?: any;
}